Add peginator_codegen crate
This splits off code generation into a library crate which can be consumed
by those who need it. In particular, it allows us to isolate `clap` (and its
large dependency tree) for just the cli utility. The resulting dependency graph
looks like this:

  ┌───────────────────┐   ┌───────────────────┐
  │                   │   │                   │
  │ peginator_codegen │◄──┤ peginator_runtime ├──┐
  │                   │   │                   │  │
  └─────────┬────┬────┘   └───────────────────┘  │
            │    │                               │
            │    └──────────────────┐            │
            ▼                       ▼            │
  ┌───────────────────┐   ┌───────────────────┐  │
  │                   │   │                   │  │
  │  peginator (cli)  │   │    buildscript    │  │
  │                   │   │                   │  │
  └──────────────┰────┘   └────┰──────────────┘  │
                 ┇             ┇                 │
                 ┇  generates  ┇                 │
                 ▼             ▼                 │
              ┌───────────────────┐              │
              │                   │              │
              │  generated code   │◄─────────────┘
              │                   │
              └───────────────────┘
