+++ Cargo.toml
@@ -14,14 +14,8 @@ keywords = ["parsing", "parser", "grammar", "ast", "peg"]
 categories = ["parsing"]
 license = "MIT"
 
-[[bin]]
-name = "peginator-compile"
-
 [dependencies]
-peginator_runtime = { version="=0.4.0", path = "runtime" }
-quote = "1.0"
-proc-macro2 = "1.0"
-anyhow = "1.0"
+peginator_codegen = { version="=0.4.0", path = "codegen" }
 clap = { version = "4.0", features = ["derive"]}
 colored = "2.0"
-crc = "3.0"
+anyhow = "1.0"
