+++ bootstrap.sh
@@ -1,6 +1,6 @@
 #/usr/bin/env sh
 
-OUTPUT="src/grammar/generated.rs"
+OUTPUT="codegen/src/grammar/generated.rs"
 
 if cargo run -- grammar.ebnf | rustfmt > $OUTPUT.new; then
     if mv -f $OUTPUT.new $OUTPUT; then
