+++ codegen/Cargo.toml
@@ -1,20 +1,20 @@
 [package]
-name = "peginator_macro"
-description = "Macros for peginator"
+name = "peginator_codegen"
+description = "PEG parser generator library for creating ASTs in Rust"
 version = "0.4.0"
 edition = "2021"
 
 authors = ["Alex Badics <admin@stickman.hu>"]
 repository = "https://github.com/badicsalex/peginator"
-documentation = "https://docs.rs/peginator_macro"
+documentation = "https://docs.rs/peginator_codegen"
 keywords = ["parsing", "parser", "grammar", "ast", "peg"]
 categories = ["parsing"]
 license = "MIT"
 
-[lib]
-proc-macro = true
-
 [dependencies]
 peginator_runtime = { version="=0.4.0", path = "../runtime" }
-peginator_codegen = { version="=0.4.0", path = "../codegen" }
-syn = { version = "1.0", default-features = false, features = ["parsing", "proc-macro"] }
+quote = "1.0"
+proc-macro2 = "1.0"
+anyhow = "1.0"
+colored = "2.0"
+crc = "3.0"
