+++ codegen/src/choice.rs
