+++ codegen/src/common.rs
