+++ codegen/src/eoi.rs
