+++ codegen/src/lib.rs
@@ -2,6 +2,15 @@
 // This file is part of peginator
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 
+//! This crate contains the code used for generating parsing code from a
+//! grammar file. Unless you are using [`Compile`] in a buildscript, you
+//! probably want to see the [peginator] crate documentation instead.
+//!
+//! [peginator]: https://docs.rs/peginator/latest/peginator
+
+pub mod grammar;
+
+mod buildscript;
 mod char_rule;
 mod choice;
 mod closure;
@@ -9,7 +18,6 @@ mod common;
 mod eoi;
 mod extern_rule;
 mod field;
-mod grammar;
 mod header;
 mod include_rule;
 mod lookahead;
@@ -19,6 +27,9 @@ mod rule;
 mod sequence;
 mod string;
 
+pub use peginator_runtime as runtime;
+
+pub use buildscript::Compile;
 pub use common::{CodegenGrammar, CodegenSettings};
 pub use header::generate_source_header;
