+++ codegen/src/string.rs
