+++ macro/README.md
@@ -6,10 +6,10 @@ doucmentation and other goodies on the main project.
 Usage:
 ```rust
 use peginator_macro::peginate;
-use peginator::PegParser;
+use peginator_runtime::PegParser;
 
 peginate!("
-@export 
+@export
 PizzaRule =
     'Pizza' 'with'
     toppings:Topping
