+++ macro/tests/simple.rs
@@ -2,8 +2,8 @@
 // This file is part of peginator
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 
-use peginator::PegParser;
 use peginator_macro::peginate;
+use peginator_runtime::PegParser;
 
 peginate!("@export Simple = c:char;");
