--- src/grammar/mod.rs
@@ -1,7 +1,0 @@
-// Copyright (C) 2022, Alex Badics
-// This file is part of peginator
-// Licensed under the MIT license. See LICENSE file in the project root for details.
-
-mod generated;
-
-pub use generated::*;
