+++ test/Cargo.toml
@@ -13,10 +13,9 @@ publish = false
 
 [dependencies]
 peginator_runtime = { version="=0.4.0", path = "../runtime" }
-peginator = { path = "../" }
 
 [build-dependencies]
-peginator = { path = "../" }
+peginator_codegen = { version="=0.4.0", path = "../codegen" }
 
 [dev-dependencies]
 ntest = "0.9"
