+++ test/build.rs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 
 fn main() {
-    peginator::buildscript::Compile::directory("src")
+    peginator_codegen::Compile::directory("src")
         .format()
         .derives(vec![
             "Debug".into(),
@@ -12,12 +12,12 @@ fn main() {
             "Eq".into(),
         ])
         .run_exit_on_error();
-    peginator::buildscript::Compile::file("src/custom_derives_empty/grammar.not_ebnf")
+    peginator_codegen::Compile::file("src/custom_derives_empty/grammar.not_ebnf")
         .format()
         .derives(vec![])
         .prefix("pub struct ImJustHereToConfuse;".into())
         .run_exit_on_error();
-    peginator::buildscript::Compile::file("src/user_defined_state/grammar.not_ebnf")
+    peginator_codegen::Compile::file("src/user_defined_state/grammar.not_ebnf")
         .format()
         .derives(vec![
             "Debug".into(),
