+++ test/src/calculator_example/mod.rs
@@ -4,7 +4,7 @@
 
 mod grammar;
 use grammar::*;
-use peginator::PegParser;
+use peginator_runtime::PegParser;
 
 trait Calculate {
     fn calculate(&self) -> i64;
