+++ test/src/skip_ws/mod.rs
@@ -4,7 +4,7 @@
 
 mod grammar;
 use grammar::*;
-use peginator::PegParser;
+use peginator_runtime::PegParser;
 
 #[test]
 fn test_string() {
